adaptive-engine/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
openai==1.30.0
python-dotenv==1.0.1
pydantic==2.7.1
